Synth set frequency calcs use integers for desired freq instead of floats (more accurate). Freq modulation still WIP.

+++ software/stm32f4/AusBubble/src/ScanAlgorithms.cpp
@@ -46,9 +46,9 @@ void ScanInit(scanSettings_t* settings)
 
 void AdvanceScan(scanSettings_t* settings)
 {
-    static uint32_t freq = settings->start;
+    static uint64_t freq = settings->start;
     static bool direction = DIR_UP;
-    uint32_t newFreq = -1;
+    uint64_t newFreq = -1;
     uint32_t random32bit = 0;
 
     switch(settings->algorithm)
